REFACTOR: init the page
