+++ src/app/about/page.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 const AboutPage = () => {
-  return <div className="h-[2000px] bg-slate-400">AboutPage</div>;
+  return <div className="post h-[2000px]">AboutPage</div>;
 };
 
 export default AboutPage;
