+++ src/components/layout/ScrollProgressBar.tsx
@@ -18,7 +18,7 @@ const ScrollProgressBar = () => {
   };
   return (
     <div
-      className="fixed top-0 z-20 h-1 w-full bg-blue-600"
+      className="fixed top-0 z-30 h-1 w-full bg-blue-600"
       style={{ transform: `translateX(${scrolled - 100}%)` }}
     />
   );
