+++ src/config/const.ts
@@ -1,5 +1,4 @@
 export const baseDomain = "https://www.d5br5.dev";
-
 export const blogName = "백승호 개발 블로그";
 export const blogDesc = "개발 지식과 경험을 공유합니다.";
 export const blogThumbnailURL = `${baseDomain}/blog_thumbnail.png`;
